requirements.txt
smolagents>=1.16.0
transformers>=4.48.0,<4.52
torch>=2.0.0
pillow>=10.0.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
numpy>=1.24.0
accelerate>=0.27.0
sentencepiece>=0.2.0
safetensors>=0.4.0
protobuf>=3.20.0
scikit-learn>=1.3.0
scipy>=1.11.0
pydantic>=2.4.0
argparse>=1.4.0
# Add smol-agent if it's a pip-installable library, otherwise note it for the user
# For now, we'll assume it's managed separately or a development paradigm
# smolagents[all]  # The smol-agent framework
gradio>=3.50.0
einops>=0.7.0
bitsandbytes>=0.40.0  # For model quantization
diffusers>=0.25.0
torchvision>=0.16.0
torchxrayvision>=1.0.0
scikit-image>=0.20.0 